Partially refactor AI, use Objective/Behaviour structure
Current design (will probably change) is that the AIs will have
Objectives mapped to Behaviours. Chossing an Objective will still
have to be custom-written for each AI. However, once an Objective
has been defined, the AI will run down the list of Behaviours
mapped to that Objective and choose the first one which is usable.

Obviously there is room for improvement.

Currently, it's partially complete only; must add in the actual
mapping (hard-coded).